split main.dart
