+++ stiknes_mobile/lib/screens/edit_note_screen.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+
+class EditNoteScreen extends StatefulWidget {
+  final Map<String, String> note;
+
+  const EditNoteScreen({super.key, required this.note});
+
+  @override
+  _EditNoteScreenState createState() => _EditNoteScreenState();
+}
+
+class _EditNoteScreenState extends State<EditNoteScreen> {
+  late TextEditingController _titleController;
+  late TextEditingController _contentController;
+
+  @override
+  void initState() {
+    super.initState();
+    _titleController = TextEditingController(text: widget.note['title']);
+    _contentController = TextEditingController(text: widget.note['content']);
+  }
+
+  @override
+  void dispose() {
+    _titleController.dispose();
+    _contentController.dispose();
+    super.dispose();
+  }
+
+  void _saveNote() {
+    ScaffoldMessenger.of(context).showSnackBar(
+      const SnackBar(content: Text('Note saved successfully!')),
+    );
+    Navigator.pop(context);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Edit Note')),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            TextField(
+              controller: _titleController,
+              decoration: const InputDecoration(labelText: 'Title'),
+            ),
+            const SizedBox(height: 10),
+            TextField(
+              controller: _contentController,
+              maxLines: 5,
+              decoration: const InputDecoration(labelText: 'Content'),
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: _saveNote,
+              child: const Text('Save'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
