+++ stiknes_mobile/lib/screens/user_info_screen.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+
+class UserInfoScreen extends StatelessWidget {
+  final String username;
+
+  const UserInfoScreen({super.key, required this.username});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('User Info')),
+      body: Center(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              const Icon(Icons.person, size: 100, color: Colors.blue),
+              const SizedBox(height: 20),
+              Text(
+                'Username: $username',
+                style: const TextStyle(fontSize: 20),
+              ),
+              const SizedBox(height: 10),
+              const Text(
+                'Email: example@example.com',
+                style: TextStyle(fontSize: 18),
+              ),
+              const SizedBox(height: 20),
+              ElevatedButton(
+                onPressed: () {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    const SnackBar(content: Text('Profile editing coming soon!')),
+                  );
+                },
+                child: const Text('Edit Profile'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
